requirements.txt
streamlit==1.18
pyzbar
numpy
pillow
pandas